Uploading my attempt at P1
Me and Eric discussed the ability to have the rectangles stored with searchable labels, so I tried to store them in a dictionary.

The whole first part of this code is a really messy attempt at iterating over all of the images in the folder. Im struggling with iterating over the cpos and cneg named files and then pulling their coordinate data. I'm able to differentiate between the files by how they are named, but then for some reason my code does not store any of the coordinate data. Let me know if you see something when looking over it.